Rename sfc TheHeader
